+++ yaml-config-updater/src/test/groovy/ru/vyarus/yaml/updater/parse/common/ModelUtilsTest.groovy
@@ -20,4 +20,26 @@ class ModelUtilsTest extends AbstractTest {
         'one' | 'one/two' | 'two'
         'one' | 'one[10]' | '[10]'
     }
+
+    def "Check property cleanup"() {
+
+        expect:
+        YamlModelUtils.cleanPropertyName(prop) == result
+
+        where:
+        prop         | result
+        'prop'       | 'prop'
+        '"prop"'     | 'prop'
+        '\'prop\''   | 'prop'
+        "smth''ng"   | "smth'ng"
+        "'smth''ng'" | "smth'ng"
+    }
+
+    def "Check incorrectly quoted property"() {
+
+        when: "property with incorrect quotes"
+        YamlModelUtils.cleanPropertyName("'name")
+        then: "error"
+        thrown(IllegalStateException)
+    }
 }