+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   - Fix properties recognition: colon must follow space or newline
     (could be not quoted string containing colon or even property name with colon)
   - Fix quoted property names support
+  - Property writing could change on update according to target file
+    (e.g. quoted property may become unquoted or opposite)
 
 ### 1.3.3 (2021-11-15)
 * [api]